refactor: Improve UI responsiveness and user experience across multiple pages
- Update routing for admin question management
- Enhance quiz creation form with improved loading and redirection
- Refactor profile page charts with better layout and responsive design
- Improve simulados, temas, and testes-previos page layouts
- Add more descriptive loading and empty states
- Optimize chart components with better formatting and mobile-friendly views
