+++ src/lib/utils.ts
@@ -14,3 +14,18 @@ export function formatDate(timestamp: number) {
     minute: '2-digit',
   });
 }
+
+// Format relative time helper
+export const formatRelativeTime = (timestamp: number): string => {
+  const now = Date.now();
+  const diffInSeconds = Math.floor((now - timestamp) / 1000);
+
+  if (diffInSeconds < 60) return 'Agora mesmo';
+  if (diffInSeconds < 3600) return `${Math.floor(diffInSeconds / 60)}m atrás`;
+  if (diffInSeconds < 86_400)
+    return `${Math.floor(diffInSeconds / 3600)}h atrás`;
+  if (diffInSeconds < 604_800)
+    return `${Math.floor(diffInSeconds / 86_400)}d atrás`;
+
+  return formatDate(timestamp);
+};
